requirements.txt
django
django_filter
django-bootstrap-v5
django-bootstrap-icons
django-widget-tweaks
django-pandas
django-htmx
beautifulsoup4
django-channels
uvicorn
channels
concurrent-log-handler==0.9.26
gitpython
pandas
websockets
factory_boy